requirements.txt
numpy
scrapeasy
wheel
RPi.GPIO
fake-rpi
pymitter